pyproject.toml
[tool.poetry]
name = "stateful-event-processing"
version = "0.1.0"
description = ""
authors = ["glyphack <sh.hooshyari@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
certifi = "^2022.6.15"
confluent-kafka = {extras = ["avro"], version = "^1.9.2"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
